requirements.txt
pandas>=2.1.0        # Data manipulation
boto3>=1.34.0        # AWS S3 client
botocore>=1.38.0     # Dependency of boto3, explicitly listed
pyyaml>=6.0          # Config YAML loader
pyspark==3.4.1       # Spark for data processing